Split core i18n migrations
